add me query and test
