+++ tests/PassportTestCase.php
@@ -1,0 +1,25 @@
+<?php
+
+namespace Tests;
+
+use App\Models\User;
+use Illuminate\Foundation\Testing\WithFaker;
+use Illuminate\Foundation\Testing\RefreshDatabase;
+use Laravel\Passport\Passport;
+use Tests\TestCase;
+
+class PassportTestCase extends TestCase
+{
+    /**
+     * @var User
+     */
+    protected $user;
+
+    public function setUp()
+    {
+        parent::setUp();
+
+        $this->user = factory(User::class)->create();
+        Passport::actingAs($this->user);
+    }
+}
